Added support for 16bpp RGB565
